+++ extra/doc/CHANGELOG.md
@@ -2,5 +2,9 @@
 
 * version 1.3.0 September 2022
 	* first release
-
+* Version 1.3.1 
+	* Added fonts tiny homespun and mednums
+	* Replaced Multibuffer with single buffer mode.
+	* Examples files reduced from 5 to 3.
+	* Added Spi baud rate parameter